Remove accidentally copied code
This patch removes accidentally copied code to `tensor_iterator.h`.
